Unvalid Move Bug Fixed
