+++ api/Comm/API.py
@@ -1,26 +1,52 @@
 from flask import Flask, request, jsonify
-from Pawn import Pawn
+from flask_cors import CORS
 from Logic import *
 import torch
 
 app = Flask(__name__)
-AI = torch.load("Model.pth")
+CORS(app)  # CORS'u etkinleştir
+AI = QLinearNet()
+AI.load_state_dict(torch.load("Model.pth", weights_only=False))
+AI.eval()
 
-@app.route('/api', methods=['POST'])
-def predict():
-    data = request.get_json(force=True)
 
-    maze = data['maze']
-    playerPawn = data['playerPawn']  # X,Y
-    opponentPawn = data['opponentPawn']  # X,Y
-    walls = data['walls']  # Active Walls
+@app.route('/update-state', methods=['POST'])
+def update_state():
+    data = request.json
+    maze, players = organise_data(data)
+    walls = convertMazeToWalls(maze)
 
-    playerPawn = Pawn("Player", playerPawn[0], playerPawn[1], maze)
-    opponentPawn = Pawn("Player", opponentPawn[0], opponentPawn[1], maze)
+    playerX = players[0]['position']['col']
+    playerY = players[0]['position']['row']
+    playerWall = players[0]['wallsLeft']
+
+    opponentX = players[1]['position']['row']
+    opponentY = players[1]['position']['row']
+    opponentWall = players[1]['wallsLeft']
+
+    playerPawn = Pawn("Player", playerX, playerY, maze)  # Red
+    playerPawn.remainingWalls = playerWall
+    opponentPawn = Pawn("Player", opponentX, opponentY, maze)  # Blue
+    opponentPawn.remainingWalls = opponentWall
+
+    playerPawn.setOpponent(opponentPawn)
+    opponentPawn.setOpponent(playerPawn)
+
     state = get_state(playerPawn, opponentPawn, walls)
-    action = get_action(AI, state)
-
+    action = get_action(AI, state, playerPawn, opponentPawn)
+    action = list(action)
+    print(action)
+    if len(action) == 2:
+        action[0] += playerX
+        action[1] += playerY
+    action[0], action[1] = action[1], action[0]
     return jsonify(action)
+
+
+@app.route('/get_string', methods=['GET'])
+def get_string():
+    data = "Merhaba, bu bir test mesajıdır!"
+    return jsonify({"message": data})
 
 
 if __name__ == '__main__':
